JAMES-3365 Move Properties into their own model class
